Started adding config values to monsters -Caden

+++ Project Riptide/Assets/Scripts/Monsters/Enemy.cs
@@ -381,7 +381,7 @@ public partial class Enemy : Physics
     /// <param name="damage">Damage projectile inflicts</param>
     /// <param name="maxLifeSpan">Max life span of projectile</param>
     /// <param name="movementPattern">Movement pattern of projectile</param>
-    protected void SpawnProjectile(Vector3 position, Vector3 size, float speed, int damage, float maxLifeSpan, MovementPattern movementPattern)
+    protected void SpawnProjectile(Vector3 position, Vector3 size, float speed, float damage, float maxLifeSpan, MovementPattern movementPattern)
     {
         GameObject temp = Instantiate(_projectile,
             transform.position + transform.TransformVector(position),
@@ -406,7 +406,7 @@ public partial class Enemy : Physics
     /// <param name="movementPattern">Movement pattern of projectile</param>
     /// <param name="launchAngle">Angle that hitbox will launch player</param>
     /// <param name="launchStrength">Strength at which player will be launched</param>
-    protected void SpawnProjectile(Vector3 position, Vector3 size, float speed, int damage, float maxLifeSpan, MovementPattern movementPattern, Vector2 launchAngle, float launchStrength)
+    protected void SpawnProjectile(Vector3 position, Vector3 size, float speed, float damage, float maxLifeSpan, MovementPattern movementPattern, Vector2 launchAngle, float launchStrength)
     {
         GameObject temp = Instantiate(_projectile,
             transform.position + transform.TransformVector(position),
